+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Roadmap
+# Road map
 
 - [ ] Nothing yet...
 
@@ -6,7 +6,7 @@ Features that have a checkmark are complete and available for
 download in the
 [nightly build](http://vsixgallery.com/extension/92e3e73b-510f-45bb-8aee-c637e83778b3/).
 
-# Changelog
+# Change log
 
 These are the changes to each version that has been released
 on the official Visual Studio extension gallery.